Add reusable Snyk docker image scan action
This adds a reusable GitHub Action that uses Snyk to scan a docker image.

The action takes a `tag` input, and requires a `snyk_token` secret to run.

It uses the `Dockerfile` located in the root folder of the application.

Example usage might look something like this:

```yml
name: Snyk Docker image scan

on:
  schedule:
    - cron: "0 5 * * *"

jobs:
  snyk:
    uses: ministryofjustice/laa-reusable-github-actions/.github/workflows/snyk.yml@main
    with:
      tag: "laa-apply-for-legalaid"
    secrets:
      snyk_token: ${{ secrets.SNYK_TOKEN }}
```